requirements.txt
accelerate==1.1.1
addict==2.4.0
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deepspeed==0.14.4
docker-pycreds==0.4.0
docopt-ng==0.9.0
einops==0.8.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
hjson==3.1.0
huggingface-hub==0.26.2
idna==3.10
jaxtyping==0.2.36
Jinja2==3.1.4
jsonpickle==4.0.0
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mmengine==0.10.5
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rotary-embedding-torch==0.8.5
sacred==0.8.6
safetensors==0.4.5
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sympy==1.13.1
taylor-series-linear-attention==0.1.12
termcolor==2.5.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.0
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.7
wrapt==1.17.0
yapf==0.43.0
